requirements.txt
Django==4.0.3
djangorestframework==3.14.0
psycopg2-binary==2.9.9
redis==5.0.1
huey==2.4.5
python-decouple==3.8
dj-database-url==1.3.0
pytest==7.4.3
pytest-django==4.7.0
